#1405 unify porosity models
