--- docs/source/models/submodels/porosity/full_reaction_driven_porosity.rst
@@ -1,7 +1,0 @@
-Full Model
-==========
-
-.. autoclass:: pybamm.porosity.Full
-    :members:
-
-
