+++ docs/source/models/submodels/porosity/index.rst
@@ -6,6 +6,5 @@ Porosity
 
   base_porosity
   constant_porosity
-  leading_reaction_driven_porosity
-  full_reaction_driven_porosity
+  reaction_driven_porosity
