--- docs/source/models/submodels/porosity/leading_reaction_driven_porosity.rst
@@ -1,8 +1,0 @@
-Leading-Order Model 
-===================
-
-.. autoclass:: pybamm.porosity.LeadingOrder
-    :members:
-
-
-
