+++ docs/source/models/submodels/porosity/reaction_driven_porosity.rst
@@ -1,0 +1,8 @@
+Reaction-driven Model 
+=====================
+
+.. autoclass:: pybamm.porosity.ReactionDriven
+    :members:
+
+
+
