+++ pybamm/models/full_battery_models/lead_acid/full.py
@@ -54,7 +54,9 @@ class Full(BaseModel):
         pybamm.citations.register("Sulzer2019physical")
 
     def set_porosity_submodel(self):
-        self.submodels["porosity"] = pybamm.porosity.Full(self.param, self.options)
+        self.submodels["porosity"] = pybamm.porosity.ReactionDriven(
+            self.param, self.options, False
+        )
 
     def set_convection_submodel(self):
         if self.options["convection"] == "none":
