+++ pybamm/models/full_battery_models/lithium_ion/spm.py
@@ -68,8 +68,8 @@ class SPM(BaseModel):
             self.options["SEI porosity change"] == "true"
             or self.options["lithium plating porosity change"] == "true"
         ):
-            self.submodels["porosity"] = pybamm.porosity.LeadingOrder(
-                self.param, self.options
+            self.submodels["porosity"] = pybamm.porosity.ReactionDriven(
+                self.param, self.options, True
             )
 
     def set_active_material_submodel(self):
