+++ pybamm/models/full_battery_models/lithium_ion/spme.py
@@ -70,8 +70,8 @@ class SPMe(BaseModel):
             self.options["SEI porosity change"] == "true"
             or self.options["lithium plating porosity change"] == "true"
         ):
-            self.submodels["porosity"] = pybamm.porosity.LeadingOrder(
-                self.param, self.options
+            self.submodels["porosity"] = pybamm.porosity.ReactionDriven(
+                self.param, self.options, True
             )
 
     def set_active_material_submodel(self):
