+++ pybamm/models/submodels/porosity/__init__.py
@@ -1,4 +1,3 @@
 from .base_porosity import BaseModel
 from .constant_porosity import Constant
-from .full_reaction_driven_porosity import Full
-from .leading_reaction_driven_porosity import LeadingOrder
+from .reaction_driven_porosity import ReactionDriven
